add implementation for gitlab service, toa for convert person to gitlabuser

+++ cms/cms-core/src/main/java/org/javatraining/integration/gitlab/api/model/GitLabUserEntity.java
@@ -220,7 +220,7 @@ public class GitLabUserEntity implements Serializable {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 }
